+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+# 1.1.17 - 29 Sep 2024
+- [#838](https://github.com/elysiajs/elysia/issues/838) invalid `onAfterResponse` typing
+- [#855](https://github.com/elysiajs/elysia/issues/855) Validation with Numeric & Number options doesn't work
+
 # 1.1.16 - 23 Sep 2024
 Bug fix:
 - separate between `createStaticHandler` and `createNativeStaticHandler` for maintainability